add classes directory for code class code that's does not extend taco...
This directory is meant for classes that pertain to the app but that do
not extend Taco. It would not go in the lib directory because we are
treating that as a directory for 3rd party or util classes that we
eventually plan to add to Packagist. This directory is not meant for
code that will be used across projects but rather a place to put per
app classes.